refactor: Split up Confirm component
